write chapter 2/c which explains the object's cultural biography
The anecdote connected to The Punished Suicide – also quoted in one of the exhibition labels – is
that the woman’s act of suicide was caused by “amorous delusion.” However, researchers were able
to discover the identity of the woman used for The Punished Suicide: her name was Eugenia Perini,
a maid living in Ognissanti.

None of the official reports issued about her death mention suicide, – they indicate it as “casual
drowning” – and researches were unable to find any news reports describing the event as suicide.1
This layer of complexity can partially be explained through the cultural biography of the object,
which is defined in the following way:

“Objects can be understood only through looking at the cultural contexts which originally produced
them and the new circumstances into which they later moved. The histories of many objects are
composed of shifts of context and perspective.”2

The contexts which are relevant for the understanding of The Punished Suicide are:

(1) changing attitudes towards suicide,

(2) the rise of Romanticism and

(3) scientific objectivity.